+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Blacklist known BAN wallets used by exchanges #80
 ## Bugs
 - Welcome page - layout issue for BSC #84
+- Missing crypto dialog not entirely visible on mobile #86
 
 # v1.2.3
 ## Improvements