+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New setup page #81
 - Display warning about blockchain fees is (un)wrapping up to 100 BAN #87
 - Blacklist known BAN wallets used by exchanges #80
+- Add link to NFT page #83
 - Add Discord & Telegram links #25
 - Add a link to GitBook documentation #47
 ## Bugs